requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bleach==2.1.2
bokeh==0.12.13
certifi==2017.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
flake8==3.5.0
Flask==0.12.2
functools32==3.2.3.post2
futures==3.2.0
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.0
jupyter-core==4.4.0
jupyterlab==0.30.6
jupyterlab-launcher==0.6.0
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
pandas==0.21.1
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.0
pew==1.1.2
pexpect==4.3.1
pickleshare==0.7.4
pipenv==9.0.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
requests==2.18.4
scandir==1.6
scipy==1.0.0
shutilwhich==1.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.13
